+++ web/src/lib/components/photos-page/asset-grid.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import AssetGridWithoutScrubber from '$lib/components/photos-page/asset-grid-without-scrubber.svelte';
+  import TimelineViewer from '$lib/components/timeline-viewer/timeline-viewer.svelte';
   import { AssetAction } from '$lib/constants';
   import type { DayGroup } from '$lib/managers/timeline-manager/day-group.svelte';
   import type { MonthGroup } from '$lib/managers/timeline-manager/month-group.svelte';
@@ -185,7 +185,7 @@
   };
 </script>
 
-<AssetGridWithoutScrubber
+<TimelineViewer
   {isSelectionMode}
   {singleSelect}
   {enableRouting}
@@ -220,4 +220,4 @@
       />
     {/if}
   {/snippet}
-</AssetGridWithoutScrubber>
+</TimelineViewer>
