+++ web/src/lib/components/timeline-viewer/skeleton.svelte
