+++ web/src/lib/components/timeline-viewer/timeline-viewer.svelte
@@ -5,9 +5,8 @@
   import AssetDateGroupSelectionAware from '$lib/components/photos-page/asset-date-group-selection-aware.svelte';
   import AssetGridActions from '$lib/components/photos-page/asset-grid-actions.svelte';
   import AssetGridAssetViewer from '$lib/components/photos-page/asset-grid-asset-viewer.svelte';
-  import Skeleton from '$lib/components/photos-page/skeleton.svelte';
+  import Skeleton from '$lib/components/timeline-viewer/skeleton.svelte';
   import { AssetAction } from '$lib/constants';
-  import type { DayGroup } from '$lib/managers/timeline-manager/day-group.svelte';
   import type { MonthGroup } from '$lib/managers/timeline-manager/month-group.svelte';
   import { TimelineManager } from '$lib/managers/timeline-manager/timeline-manager.svelte';
   import type { TimelineAsset } from '$lib/managers/timeline-manager/types';
@@ -47,18 +46,6 @@
     children?: Snippet;
     empty?: Snippet;
     handleTimelineScroll?: () => void;
-    customLayout?: Snippet<[TimelineAsset]>;
-    onThumbnailClick?: (
-      asset: TimelineAsset,
-      timelineManager: TimelineManager,
-      dayGroup: DayGroup,
-      onClick: (
-        timelineManager: TimelineManager,
-        assets: TimelineAsset[],
-        groupTitle: string,
-        asset: TimelineAsset,
-      ) => void,
-    ) => void;
   }
 
   let {
@@ -80,8 +67,6 @@
     empty,
     header,
     handleTimelineScroll = () => {},
-    customLayout,
-    onThumbnailClick,
   }: Props = $props();
 
   let { isViewing: showAssetViewer, gridScrollTarget } = assetViewingStore;
@@ -342,8 +327,6 @@
             {onSelect}
             {onScrollToTop}
             {onScrollCompensation}
-            {customLayout}
-            {onThumbnailClick}
           />
         </div>
       {/if}
